+++ test/IOEither.ts
@@ -1,13 +1,12 @@
 import * as assert from 'assert'
 import { sequenceT } from '../src/Apply'
 import * as E from '../src/Either'
-import { pipe } from '../src/function'
+import { pipe, Predicate } from '../src/function'
 import * as I from '../src/IO'
 import * as _ from '../src/IOEither'
 import * as A from '../src/ReadonlyArray'
 import { monoidString } from '../src/Monoid'
 import { none, some } from '../src/Option'
-import { pipeable } from '../src/pipeable'
 import { semigroupSum } from '../src/Semigroup'
 
 describe('IOEither', () => {
@@ -330,7 +329,10 @@ describe('IOEither', () => {
 
   describe('getFilterable', () => {
     const F_ = _.getFilterable(A.getMonoid<string>())
-    const { filter } = pipeable(F_)
+    const filter: <A>(
+      predicate: Predicate<A>
+    ) => (fa: _.IOEither<ReadonlyArray<string>, A>) => _.IOEither<ReadonlyArray<string>, A> = (predicate) => (fa) =>
+      F_.filter(fa, predicate)
 
     it('filter', async () => {
       const r1 = pipe(